Add Settlement Deadline (#13)
With #11, closes #10.

This PR adds a signed **block** deadline parameter to the settlement
trampoline contract. The motivation behind this change is to make it
easier for solvers to manage in-flight trampolined settlements: if a
solver sees a certain block height and their settlement has not yet been
mined, they know that they don't need to execute a `cancelCurrentNonce`
transaction, and can reuse the previous nonce (as the signed settlement
is no longer valid). Additionally, this fits nicely into the protocol as
we want solvers to "time out" execution at some point anyway (we
currently implement this by sending "cancellation transactions" from EOA
solvers).

Note that I decided to use a block deadline instead of a time based one.
The idea here is that we want to transistion the protocol to be more
block-centric, and I think this is a step in the right direction.
Furthermore, for PoS, block number and block timestamp are largely
equivalent for deadlines anyway.

### Test Plan

Added unit test.