Add Additional Invalid Nonce Error Context (#14)
This PR adds additional error context to the `InvalidNonce` error so
that it includes the specified nonce and the actual current nonce.

The motivation behind this change is to make debugging reverts easier.
Note that this adds a meaningless 5 additional gas to the happy path,
and so is very low cost for additional debuggability.

### Test Plan

Existing test now checks for error data.